Fixed a Mentionable Data bug
The mentionable cache object was not made with guild id's in mind, so if another server would request the object, it would get the roles from another server.
I made the cache object have keys that are the guild id's so that i filters through the cache for each guild.