Update data.ts
WIP, cant do this rn, its freaking me out. but im planning to instead of the raw object for data, make that the actual class object instance, but not or the input, but only as the result, so that what ever is stored in the client collections, like commands, if you do client.commands['command'].data you would get a commandObject and not the raw data that was input into it. this is to preserve the functions that are included in there.