AGONY
i cant even begin to describe how annoying this is... but for the most important things, i want to command interaction data to contain the raw content in object form, and also the data in class for to access the functions it contains. i got that to work in data.ts now but i need to implement a new way of storing the command in the client so its by default a class and not just the raw object.